+++ modules/bank/types.go
@@ -6,8 +6,8 @@ import (
 	"fmt"
 	"regexp"
 
+	json2 "github.com/irisnet/irishub-sdk-go/tools/json"
 	"github.com/irisnet/irishub-sdk-go/types"
-	json2 "github.com/irisnet/irishub-sdk-go/utils/json"
 )
 
 const memoRegexpLengthLimit = 50
