+++ modules/oracle/types.go
@@ -4,8 +4,8 @@ import (
 	"errors"
 	"strings"
 
+	"github.com/irisnet/irishub-sdk-go/tools/json"
 	sdk "github.com/irisnet/irishub-sdk-go/types"
-	"github.com/irisnet/irishub-sdk-go/utils/json"
 )
 
 var (
