+++ tools/bech32/bech32.go
