+++ tools/commom.go
@@ -1,4 +1,4 @@
-package utils
+package tools
 
 import "crypto/rand"
