+++ tools/uuid/codec.go
