+++ tools/uuid/generator.go
